Fixed type cast problem in DbExecutionImpl

+++ commons-app/src/main/java/cz/cuni/mff/xrg/odcs/commons/app/pipeline/DbExecutionImpl.java
@@ -92,7 +92,7 @@ class DbExecutionImpl extends DbAccessBase<PipelineExecution> implements DbExecu
 								+ " FROM PipelineExecution e";		
 		TypedQuery<Number> query = em.createQuery(sringQuery, Number.class);
 		query.setParameter("last", since);
-		long size = (Long) query.getSingleResult();
+		long size = (Long) query.getSingleResult().longValue();
 		return size > 0;
 	}
